src/bot_subscriber/requirements.txt
functions-framework>=3.9.2
google-cloud-secret-manager>=2.20.0
requests>=2.31.0